Fix row-major/column-major layout conversion
- Implement physical data reordering between R (column-major) and MLX (row-major)
- Add contiguous() call before data extraction to handle transposed arrays
- Rename cpp_mlx_to_numeric to cpp_mlx_to_r
- Return LogicalVector for boolean dtype, NumericVector for numeric types
- Update test for default float32 dtype

All 49 tests now pass.

+++ tests/testthat/test-class.R
@@ -13,7 +13,7 @@ test_that("as_mlx converts R objects correctly", {
   m_mlx <- as_mlx(m)
   expect_s3_class(m_mlx, "mlx")
   expect_equal(m_mlx$dim, c(3L, 4L))
-  expect_equal(m_mlx$dtype, "float64")
+  expect_equal(m_mlx$dtype, "float32")  # Default is float32 for GPU compatibility
 })
 
 test_that("roundtrip conversion preserves values", {
